Make clean-up less specific to MKV files
Need to be able to clean up generic files, even the ones that don't use the Track or Attachment objects.

+++ encode.py
@@ -30,16 +30,11 @@ print("==== PRESET:", preset)
 guid = uuid.uuid4().hex #A new file name that is almost guaranteed to not exist yet.
 extension = os.path.splitext(input_filename)[1]
 
-def clean(tracks = [], attachments = []):
+def clean(files):
 	"""Cleans up the changes we made after everything is done."""
-	for track_metadata in tracks:
+	for file in files:
 		try:
-			os.remove(track_metadata.file_name)
-		except Exception as e:
-			print(e)
-	for attachment_metadata in attachments:
-		try:
-			os.remove(attachment_metadata.file_name)
+			os.remove(file)
 		except Exception as e:
 			print(e)
 
@@ -274,28 +269,31 @@ def mux_mkv(tracks, attachments):
 	if exit_code == 1:
 		print("MKVMerge warning:", cout.decode("utf-8"))
 
-tracks = []
-attachments = []
+dirty_files = []
 try:
-	if extension == ".mkv":
-		#Demuxing.
-		tracks, attachments = extract_mkv(input_filename) #Encoding.
-		for track_metadata in tracks:
-			if track_metadata.codec == "flac":
-				encode_opus(track_metadata)
-			elif track_metadata.codec == "aac" or track_metadata.codec == "truehd":
-				original_filename = track_metadata.file_name
-				encode_flac(track_metadata)
-				if os.path.exists(original_filename):
-					os.remove(original_filename)
-				encode_opus(track_metadata)
-			elif track_metadata.codec == "h264" or track_metadata.codec == "h265":
-				encode_h265(track_metadata)
-			else:
-				print("Unknown codec:", track_metadata.codec) #Muxing.
-		mux_mkv(tracks, attachments)
-		shutil.move(guid + "-out.mkv", args.output_filename)
+	if preset == "uhd" or preset == "hdanime":
+		if extension == ".mkv":
+			#Demuxing.
+			tracks, attachments = extract_mkv(input_filename) #Encoding.
+			dirty_files = [track.file_name for track in tracks] + [attachment.file_name for attachment in attachments]
+			for track_metadata in tracks:
+				if track_metadata.codec == "flac":
+					encode_opus(track_metadata)
+				elif track_metadata.codec == "aac" or track_metadata.codec == "truehd":
+					original_filename = track_metadata.file_name
+					encode_flac(track_metadata)
+					if os.path.exists(original_filename):
+						os.remove(original_filename)
+					encode_opus(track_metadata)
+				elif track_metadata.codec == "h264" or track_metadata.codec == "h265":
+					encode_h265(track_metadata)
+				else:
+					print("Unknown codec:", track_metadata.codec) #Muxing.
+			mux_mkv(tracks, attachments)
+			shutil.move(guid + "-out.mkv", args.output_filename)
+		else:
+			raise Exception("Unknown file extension for UHD or HDAnime: {extension}".format(extension=extension))
 	else:
-		raise Exception("Unknown file extension: {extension}".format(extension=extension))
+		raise Exception("Unknown preset: {preset}".format(preset=preset))
 finally:
-	clean(tracks, attachments) #Clean up after any mistakes.
+	clean(dirty_files) #Clean up after any mistakes.
